Deduplicate parts of validator onboarding scripts
