+++ scripts/verify_validator_submissions.py
@@ -6,24 +6,7 @@
 """
 This script verifies details provided by validators that are onboarding.
 
-Takes tab-separated values on stdin with the following columns:
-
- * Timestamp
- * Email address
- * Validator name
- * Keybase account name
- * Vote account
- * (Withdraw authority check, constant)
- * (Commission percentage check, constant)
- * (Vote promise check, constant)
- * stSOL account
- * (stSOL account mint check, constant)
- * (Added public key to keybase check)
- * Identity name as published on identity account
- * (unused column, optional)
- * Maintainer address (if applicable; optional)
-
-This script then verifies:
+Takes tab-separated form responses on stdin. This script then verifies:
 
  * The vote account withdraw authority.
  * The vote account commission.
@@ -43,18 +26,15 @@ the tsv file manually to confirm that no weird Keybase usernames etc. are in the
 
 import json
 import subprocess
-import sys
 
 from typing import Any, Dict, Iterable, NamedTuple, Optional
+from validator_onboarding import Address, ValidatorResponse, iter_rows_from_stdin
 
 
 SOLIDO_AUTHORIZED_WITHDAWER = 'GgrQiJ8s2pfHsfMbEFtNcejnzLegzZ16c9XtJ2X2FpuF'
 ST_SOL_MINT = '7dHbWXmci3dT8UFYWYZweBLXgycu7Y3iL6trKn1Y7ARj'
 VOTE_PROGRAM = 'Vote111111111111111111111111111111111111111'
 SPL_TOKEN_PROGRAM = 'TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA'
-
-# A type alias to make some things more self-documenting.
-Address = str
 
 
 def solana(*args: str) -> Any:
@@ -175,161 +155,136 @@ def print_warn(message: str) -> None:
     print_color(f'WARN  {message}', ansi_color_code='33')
 
 
-class ValidatorResponse(NamedTuple):
-    timestamp: str
-    email: str
-    validator_name: str
-    keybase_username: str
-    vote_account_address: Address
-    withdraw_authority_check: Address
-    commission_check: str
-    will_vote_check: str
-    st_sol_account_address: Address
-    st_sol_mint_check: Address
-    added_to_keybase_check: str
-    identity_name: str
-    unused: str = ''
-    maintainer_address: Address = ''
-
-    def get_vote_account(self) -> Optional[VoteAccount]:
-        try:
-            result = solana(
-                'vote-account', '--output', 'json', self.vote_account_address
-            )
-            return VoteAccount(
-                validator_identity_address=result['validatorIdentity'],
-                authorized_withdrawer=result['authorizedWithdrawer'],
-                commission=result['commission'],
-                num_votes=len(result['votes']),
-            )
-
-        except subprocess.CalledProcessError:
-            return None
-
-    def check(
-        self,
-        validators_by_identity: Dict[Address, ValidatorInfo],
-        vote_accounts: Dict[Address, str],
-        identity_accounts: Dict[Address, str],
-        st_sol_accounts: Dict[Address, str],
-    ) -> None:
-        print('\n' + self.validator_name)
-        vote_account = self.get_vote_account()
-
-        if vote_account is not None:
-            print_ok('Vote account address holds a vote account.')
-        else:
-            print_error('Vote account address does not hold a vote account.')
-            return
-
-        if vote_account.authorized_withdrawer == SOLIDO_AUTHORIZED_WITHDAWER:
-            print_ok('Authorized withdrawer set to Solido.')
-        else:
-            print_error('Wrong authorized withdrawer.')
-
-        if vote_account.num_votes > 0:
-            print_ok('Vote account has votes.')
-        else:
-            print_warn('Vote account has not voted yet.')
-
-        if vote_account.commission == 100:
-            print_ok('Vote account commission is 100%.')
-        else:
-            print_error('Vote account commission is not 100%.')
-
-        validator_info = validators_by_identity.get(
-            vote_account.validator_identity_address
+def get_vote_account(self: ValidatorResponse) -> Optional[VoteAccount]:
+    try:
+        result = solana(
+            'vote-account', '--output', 'json', self.vote_account_address
         )
-        if validator_info is None:
-            print_error('Validator identity does not exist.')
-            return
-        else:
-            print_ok('Validator identity exists.')
-
-        if validator_info.keybase_username == self.keybase_username:
-            print_ok('Keybase username in form matches username in identity account.')
-        else:
-            print_error('Keybase username in identity account does not match the form.')
-
-        if validator_info.name is not None and validator_info.name.startswith(
-            'Lido / '
-        ):
-            print_ok('Validator identity name starts with "Lido / ".')
-        else:
-            print_error('Validator identity name does not start with "Lido / ".')
-
-        if validator_info.name == self.identity_name:
-            print_ok('Name in identity account matches name in form.')
-        else:
-            print_error('Name in identity account does not mach name in form.')
-
-        if check_keybase_has_identity_address(
-            self.keybase_username, vote_account.validator_identity_address
-        ):
-            print_ok(
-                f'Validator identity public key is on Keybase under {self.keybase_username}.'
-            )
-        else:
-            print_error('Could not verify validator identity through Keybase.')
-
-        token_account = get_token_account(self.st_sol_account_address)
-        if token_account is not None:
-            print_ok('Fee account exists.')
-        else:
-            print_error(f'Fee account {self.st_sol_account_address} does not exist.')
-            return
-
-        if token_account.state == "initialized":
-            print_ok('Token account is in an initialized state.')
-        else:
-            print_error('Token account is not initialized state.')
-
-        if token_account.mint_address == ST_SOL_MINT:
-            print_ok('Fee account is an stSOL account (it has the correct mint).')
-        else:
-            print_error('Fee account is not an stSOL account, the mint is wrong.')
-
-        name = vote_accounts.setdefault(self.vote_account_address, self.validator_name)
-        if name != self.validator_name:
-            print_error(f'Vote account is already in use by {name}.')
-        else:
-            print_ok('Vote account address is unique among responses seen so far.')
-
-        name = identity_accounts.setdefault(
-            vote_account.validator_identity_address, self.validator_name
+        return VoteAccount(
+            validator_identity_address=result['validatorIdentity'],
+            authorized_withdrawer=result['authorizedWithdrawer'],
+            commission=result['commission'],
+            num_votes=len(result['votes']),
         )
-        if name != self.validator_name:
-            print_error(f'Identity account is already in use by {name}.')
-        else:
-            print_ok('Identity account is unique among responses seen so far.')
 
-        name = st_sol_accounts.setdefault(
-            self.st_sol_account_address, self.validator_name
+    except subprocess.CalledProcessError:
+        return None
+
+
+def check_validator_response(
+    self: ValidatorResponse,
+    validators_by_identity: Dict[Address, ValidatorInfo],
+    vote_accounts: Dict[Address, str],
+    identity_accounts: Dict[Address, str],
+    st_sol_accounts: Dict[Address, str],
+) -> None:
+    print('\n' + self.validator_name)
+    vote_account = get_vote_account(self)
+
+    if vote_account is not None:
+        print_ok('Vote account address holds a vote account.')
+    else:
+        print_error('Vote account address does not hold a vote account.')
+        return
+
+    if vote_account.authorized_withdrawer == SOLIDO_AUTHORIZED_WITHDAWER:
+        print_ok('Authorized withdrawer set to Solido.')
+    else:
+        print_error('Wrong authorized withdrawer.')
+
+    if vote_account.num_votes > 0:
+        print_ok('Vote account has votes.')
+    else:
+        print_warn('Vote account has not voted yet.')
+
+    if vote_account.commission == 100:
+        print_ok('Vote account commission is 100%.')
+    else:
+        print_error('Vote account commission is not 100%.')
+
+    validator_info = validators_by_identity.get(
+        vote_account.validator_identity_address
+    )
+    if validator_info is None:
+        print_error('Validator identity does not exist.')
+        return
+    else:
+        print_ok('Validator identity exists.')
+
+    if validator_info.keybase_username == self.keybase_username:
+        print_ok('Keybase username in form matches username in identity account.')
+    else:
+        print_error('Keybase username in identity account does not match the form.')
+
+    if validator_info.name is not None and validator_info.name.startswith(
+        'Lido / '
+    ):
+        print_ok('Validator identity name starts with "Lido / ".')
+    else:
+        print_error('Validator identity name does not start with "Lido / ".')
+
+    if validator_info.name == self.identity_name:
+        print_ok('Name in identity account matches name in form.')
+    else:
+        print_error('Name in identity account does not mach name in form.')
+
+    if check_keybase_has_identity_address(
+        self.keybase_username, vote_account.validator_identity_address
+    ):
+        print_ok(
+            f'Validator identity public key is on Keybase under {self.keybase_username}.'
         )
-        if name != self.validator_name:
-            print_error(f'Fee stSOL account is already in use by {name}.')
-        else:
-            print_ok('Fee stSOL account is unique among responses seen so far.')
+    else:
+        print_error('Could not verify validator identity through Keybase.')
 
-        if get_account_owner(self.vote_account_address) == VOTE_PROGRAM:
-            print_ok('Vote account is owned by the vote program.')
-        else:
-            print_error('Vote account is not owned by the vote program.')
+    token_account = get_token_account(self.st_sol_account_address)
+    if token_account is not None:
+        print_ok('Fee account exists.')
+    else:
+        print_error(f'Fee account {self.st_sol_account_address} does not exist.')
+        return
 
-        if get_account_owner(self.st_sol_account_address) == SPL_TOKEN_PROGRAM:
-            print_ok('Fee stSOL account is owned by the SPL token program.')
-        else:
-            print_error('Fee stSOL account is not owned by the SPL token program.')
+    if token_account.state == "initialized":
+        print_ok('Token account is in an initialized state.')
+    else:
+        print_error('Token account is not initialized state.')
 
+    if token_account.mint_address == ST_SOL_MINT:
+        print_ok('Fee account is an stSOL account (it has the correct mint).')
+    else:
+        print_error('Fee account is not an stSOL account, the mint is wrong.')
 
-def iter_rows_from_stdin() -> Iterable[ValidatorResponse]:
-    """
-    Yield rows from stdin, including header, excluding blank lines.
-    """
-    for line in sys.stdin:
-        if line.strip() == '':
-            continue
-        yield ValidatorResponse(*line.strip().split('\t'))
+    name = vote_accounts.setdefault(self.vote_account_address, self.validator_name)
+    if name != self.validator_name:
+        print_error(f'Vote account is already in use by {name}.')
+    else:
+        print_ok('Vote account address is unique among responses seen so far.')
+
+    name = identity_accounts.setdefault(
+        vote_account.validator_identity_address, self.validator_name
+    )
+    if name != self.validator_name:
+        print_error(f'Identity account is already in use by {name}.')
+    else:
+        print_ok('Identity account is unique among responses seen so far.')
+
+    name = st_sol_accounts.setdefault(
+        self.st_sol_account_address, self.validator_name
+    )
+    if name != self.validator_name:
+        print_error(f'Fee stSOL account is already in use by {name}.')
+    else:
+        print_ok('Fee stSOL account is unique among responses seen so far.')
+
+    if get_account_owner(self.vote_account_address) == VOTE_PROGRAM:
+        print_ok('Vote account is owned by the vote program.')
+    else:
+        print_error('Vote account is not owned by the vote program.')
+
+    if get_account_owner(self.st_sol_account_address) == SPL_TOKEN_PROGRAM:
+        print_ok('Fee stSOL account is owned by the SPL token program.')
+    else:
+        print_error('Fee stSOL account is not owned by the SPL token program.')
 
 
 def main() -> None:
@@ -346,12 +301,9 @@ def main() -> None:
     identity_accounts: Dict[str, str] = {}
     st_sol_accounts: Dict[str, str] = {}
 
-    for row in iter_rows_from_stdin():
-        if row.timestamp == 'Timestamp':
-            # This is the header row, skip over it.
-            continue
-
-        row.check(
+    for response in iter_rows_from_stdin():
+        check_validator_response(
+            response,
             validators_by_identity,
             vote_accounts,
             identity_accounts,
